requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
celery==5.4.0
redis==4.6.0
alembic==1.15.1
SQLAlchemy==2.0.38
pydantic==2.10.6
python-dotenv==1.0.1
requests==2.32.3
worker==2018.4.1
flower==2.0.1
fastapi-cache2==0.2.2
logcc==2018.3.1
aiomysql==0.2.0
pydantic==2.10.6
email_validator==2.2.0
pydantic[email]==2.10.6
PyJWT==2.10.1
passlib==1.7.4
SQLAlchemy==2.0.38
Jinja2==3.1.6
asyncpg==0.30.0
